# server/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {},
		"securitySchemes": {}
	},
	"info": {
		"title": "server",
		"version": "0.6.3",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "domi"
		}
	},
	"paths": {
		"/api/gifsearch/tenor": {
			"get": {
				"operationId": "TenorSearch",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"list": {
											"items": {
												"properties": {
													"height2Width": {
														"type": "number",
														"format": "double"
													},
													"url": {
														"type": "string"
													},
													"mediumUrl": {
														"type": "string"
													},
													"tinyUrl": {
														"type": "string"
													},
													"nanoUrl": {
														"type": "string"
													},
													"id": {
														"type": "string"
													}
												},
												"required": [
													"height2Width",
													"url",
													"mediumUrl",
													"tinyUrl",
													"nanoUrl",
													"id"
												],
												"type": "object"
											},
											"type": "array"
										},
										"next": {
											"type": "string"
										}
									},
									"required": [
										"list",
										"next"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "str",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "next",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/gifsearch/klipy": {
			"get": {
				"operationId": "KlipySearch",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"list": {
											"items": {
												"properties": {
													"height2Width": {
														"type": "number",
														"format": "double"
													},
													"url": {
														"type": "string"
													},
													"mediumUrl": {
														"type": "string"
													},
													"tinyUrl": {
														"type": "string"
													},
													"nanoUrl": {
														"type": "string"
													},
													"id": {
														"type": "string"
													}
												},
												"required": [
													"height2Width",
													"url",
													"mediumUrl",
													"tinyUrl",
													"nanoUrl",
													"id"
												],
												"type": "object"
											},
											"type": "array"
										},
										"next": {
											"type": "string"
										}
									},
									"required": [
										"list",
										"next"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "str",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "next",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:3000/"
		}
	]
}